Pass the ref data feed from controller to the injestion service
