+++ test/controllers/ingestion/ReferenceDataListControllerSpec.scala
@@ -52,7 +52,7 @@ class ReferenceDataListControllerSpec extends SpecBase with GuiceOneAppPerSuite 
 
     "returns ACCEPTED when the data has been decompressed, validated and processed" in {
       when(mockReferenceDataService.validateAndDecompress(any(), any())).thenReturn(Right(testJson))
-      when(mockReferenceDataService.insert(any())).thenReturn(Future.successful(DataProcessingSuccessful))
+      when(mockReferenceDataService.insert(any(), any())).thenReturn(Future.successful(DataProcessingSuccessful))
 
       val result = route(app, fakeRequest).value
 
@@ -61,7 +61,7 @@ class ReferenceDataListControllerSpec extends SpecBase with GuiceOneAppPerSuite 
 
     "returns Bad Request when a decompression or validation error occurs" in {
       when(mockReferenceDataService.validateAndDecompress(any(), any())).thenReturn(Left(OtherError("error")))
-      when(mockReferenceDataService.insert(any())).thenReturn(Future.successful(DataProcessingSuccessful))
+      when(mockReferenceDataService.insert(any(), any())).thenReturn(Future.successful(DataProcessingSuccessful))
 
       val result = route(app, fakeRequest).value
 
@@ -70,7 +70,7 @@ class ReferenceDataListControllerSpec extends SpecBase with GuiceOneAppPerSuite 
 
     "returns with an Internal Server Error when the has been validated but data was not processed successfully" in {
       when(mockReferenceDataService.validateAndDecompress(any(), any())).thenReturn(Right(testJson))
-      when(mockReferenceDataService.insert(any())).thenReturn(Future.successful(DataProcessingFailed))
+      when(mockReferenceDataService.insert(any(), any())).thenReturn(Future.successful(DataProcessingFailed))
 
       val result = route(app, fakeRequest).value
